requirements.txt
fastapi 
psycopg2 
jinja2 
uvicorn 
python-dotenv
python-multipart
